requirements.txt
Pillow
qiskit
matplotlib
pylatexenc
